Update ifacialmocap_v2.py
Modify so that the data sent from the iFacialMocap app can be supported by both v1 and v2. Modify to communicate with the Microsoft Store version of the iFacialMocap desktop app, etc. By pressing the "Start Capture" button with the Capture Device IP set to "127.0.0.1", you can start the capture.